--- src/scipp/spatial/transform/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from . import affine, rotation, scaling, translation
